requirement.txt
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
babel==2.16.0
binaryornot==0.4.4
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
cryptography==43.0.3
cssselect2==0.7.0
daphne==4.1.2
dj-rest-auth==7.0.1
Django==5.1.1
django-browser-reload==1.15.0
django-cors-headers==4.6.0
django-environ==0.12.0
django-filter==24.3
django-getenv==1.3.2
django-simple-history==3.7.0
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
environ==1.0
et_xmlfile==2.0.0
fonttools==4.54.1
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.4
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysqlclient==2.2.4
openpyxl==3.1.5
oscrypto==1.3.0
pillow==10.4.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.10.1
pyOpenSSL==25.0.0
pypdf==5.1.0
pyphen==0.15.0
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
rich==13.8.1
service-identity==24.2.0
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
weasyprint==63.0
webencodings==0.5.1
xhtml2pdf==0.2.16
zope.interface==7.2
zopfli==0.2.3.post1
